move postgres-specific functions to their proper place
