+++ src/dream/QueryDriver/Base.ts
@@ -1,7 +1,12 @@
 import { DeleteQueryBuilder, SelectQueryBuilder, UpdateQueryBuilder } from 'kysely'
 import Dream from '../../Dream.js'
 import { AssociationStatement } from '../../types/associations/shared.js'
-import { DreamColumnNames, DreamConstructorType, DreamTableSchema } from '../../types/dream.js'
+import {
+  DreamColumnNames,
+  DreamConstructorType,
+  DreamTableSchema,
+  PrimaryKeyType,
+} from '../../types/dream.js'
 import {
   PreloadedDreamsAndWhatTheyPointTo,
   QueryToKyselyDBType,
@@ -147,6 +152,51 @@ export default class QueryDriverBase<DreamInstance extends Dream> {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
   >(type: QueryType): ToKyselyReturnType {
     throw new Error('implement toKysely in child class (if it makes sense)')
+  }
+
+  /**
+   * Builds a new DreamTransaction instance, provides
+   * the instance to the provided callback.
+   *
+   * ```ts
+   * await ApplicationModel.transaction(async txn => {
+   *   const user = await User.txn(txn).create({ email: 'how@yadoin' })
+   *   await Pet.txn(txn).create({ user })
+   * })
+   * ```
+   *
+   * @param callback - A callback function to call. The transaction provided to the callback can be passed to subsequent database calls within the transaction callback
+   * @returns void
+   */
+  // eslint-disable-next-line @typescript-eslint/require-await
+  public static async transaction<
+    DreamInstance extends Dream,
+    CB extends (txn: DreamTransaction<DreamInstance>) => unknown,
+    RetType extends ReturnType<CB>,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  >(dreamInstance: DreamInstance, callback: CB): Promise<RetType> {
+    throw new Error('implement transaction in child class')
+  }
+
+  /**
+   * @internal
+   *
+   * returns the foreign key type based on the primary key received.
+   * gives the driver the opportunity to switch i.e. bigserial to bigint.
+   */
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  public static foreignKeyTypeFromPrimaryKey(primaryKey: PrimaryKeyType) {
+    throw new Error('implement foreignKeyTypeFromPrimaryKey in child class')
+  }
+
+  /**
+   * @internal
+   *
+   * used to return the computed primary key type based
+   * on the primaryKeyType set in the DreamApp class.
+   */
+  public static primaryKeyType() {
+    throw new Error('implement primaryKeyType in child class')
   }
 
   /**
