+++ src/dream/QueryDriver/helpers/kysely/foreignKeyTypeFromPrimaryKey.ts
@@ -1,4 +1,4 @@
-import { PrimaryKeyType } from '../../types/dream.js'
+import { PrimaryKeyType } from '../../../../types/dream.js'
 
 export default function foreignKeyTypeFromPrimaryKey(primaryKey: PrimaryKeyType) {
   switch (primaryKey) {
